Pin base images to RHEL minor version
After thinking about it for a while, I think we should explicitly stick
to a given RHEL minor release as our FROM image, and manage an upgrade
to the new minor release when available, bumping our own minor version
at the same time. Rationale: a given RPM might have a function or ABI
change in a RHEL minor release (I believe), which could cause behaviour
changes in our images, so we should gate those changes with a minor
version bump ourselves.

This is necessary for the RHEL7 bases too, as current minor is 7.8 and
there's a 7.9 in the pipeline.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>