Stop referencing Image :latest tags
Substitute every reference to an OpenJDK Image (not ImageStream) :latest
tag with the current latest major.minor version instead.

ImageStream tag definitions are not modified, so where ":latest"
ImageStream tags already exist, they should continue to do so, just
referring to the back-end Images by different tags.

For OpenShift end-users, there should be no impact, as the ImageStreams
themselves retain the same tags pointing at (ultimately) the same
images.

This is in preparation for moving the OpenJDK images to "multi-stream",
a consequence of which is the image ":latest" tags will stop being
published in the Container Catalog. I'm pushing it now, to give it a
chance to percolate through OpenShift installations ahead of that
switch.

This does not impact the OpenJ9 images.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>