[OPENJDK-3655] remove or adjust chmod calls for installed scripts
Many of the shell scripts we copy into the image are shell "libraries"
which are intended to be sourced by other shell scripts and not directly
executed themselves. It's confusing to install them with +x permissions.

The scripts which are intended to be executed were set with owner or
group execute permission but "other" without. This was a problem for
some users of the containers in a non-OpenShift context, and there was
no good reason for it, so set those +x for other as well.